+++ itAssets-endpoint/src/main/java/com/anabatic/itAssets/endpoint/Response/GetAllCandidateResponse.java
@@ -4,7 +4,7 @@ public class GetAllCandidateResponse {
 	private Long id;
 	private String name;
 	private String email;
-	private Integer phoneNo; 
+	private Long phoneNo; 
 	private String skills;
 	private Float experience;
 	private Integer status;
@@ -42,10 +42,11 @@ public class GetAllCandidateResponse {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public Integer getPhoneNo() {
+
+	public Long getPhoneNo() {
 		return phoneNo;
 	}
-	public void setPhoneNo(Integer phoneNo) {
+	public void setPhoneNo(Long phoneNo) {
 		this.phoneNo = phoneNo;
 	}
 	public String getSkills() {
